// src/utils/interfaces.ts
export interface detailedMovieInfo {
  Title: string;
  Year: string;
  Rated: string;
  Released: string;
  Runtime: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Language: string;
  Country: string;
  Awards: string;
  Poster: string;
  imdbID: string;
  imdbRating: string;
}

export interface briefMovieInfo {
  Title: string;
  Year: string;
  Poster: string;
  imdbID: string;
}